extract `activeFilters` to a composable
