+++ components/collection/activity/FilterFloatButton.vue
@@ -10,18 +10,14 @@
 <script setup lang="ts">
 import { NeoButton } from '@kodadot1/brick'
 import { usePreferencesStore } from '@/stores/preferences'
+import useActiveRouterFilters from '@/composables/useActiveRouterFilters'
 
-const route = useRoute()
 const preferencesStore = usePreferencesStore()
+const activeFilters = useActiveRouterFilters()
 
-const numOfActiveFilters = computed(() => {
-  const query = { ...route.query, redesign: undefined }
-  const activeFilters = Object.entries(query).filter(
-    ([key, value]) => (key === 'search' && Boolean(value)) || value === 'true'
-  )
-
-  return activeFilters.length
-})
+const numOfActiveFilters = computed(
+  () => Object.keys(activeFilters.value).length
+)
 const openMobileFilters = () => preferencesStore.setMobileFilterCollapse(true)
 </script>
